# shared/api-contract/openapi/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hotels": {
      "post": {
        "operationId": "HotelsController_createHotel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tenantId": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 160
                  },
                  "timezone": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "default": "Asia/Saigon"
                  },
                  "brandSettings": {
                    "type": "object",
                    "additionalProperties": true
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đã tạo khách sạn",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 201
                    },
                    "error": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo khách sạn thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "tenantId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "brandSettings": {
                          "type": "object",
                          "additionalProperties": true,
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "tenant": {
                          "type": "object",
                          "nullable": true,
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "code": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          },
                          "required": [
                            "id",
                            "code",
                            "name",
                            "createdAt",
                            "updatedAt"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "tenantId",
                        "name",
                        "code",
                        "timezone",
                        "status",
                        "brandSettings",
                        "createdAt",
                        "updatedAt",
                        "tenant"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "error",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      },
      "get": {
        "operationId": "HotelsController_listHotels",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant scope chính; tenantId query chỉ là fallback tương thích tạm thời.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đã lấy danh sách khách sạn",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "error": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách khách sạn thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "tenantId": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "timezone": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              },
                              "brandSettings": {
                                "type": "object",
                                "additionalProperties": true,
                                "nullable": true
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "tenant": {
                                "type": "object",
                                "nullable": true,
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "code": {
                                    "type": "string"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  },
                                  "updatedAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": [
                                  "id",
                                  "code",
                                  "name",
                                  "createdAt",
                                  "updatedAt"
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "tenantId",
                              "name",
                              "code",
                              "timezone",
                              "status",
                              "brandSettings",
                              "createdAt",
                              "updatedAt",
                              "tenant"
                            ]
                          }
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "items"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "error",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      }
    },
...
    "/hotels/{hotelId}/requests": {
      "get": {
        "operationId": "HotelRequestsController_listRequests",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NORMAL",
                "URGENT"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đã lấy danh sách yêu cầu của khách"
          }
        },
        "tags": [
          "hotel-requests"
        ]
      }
    },
    "/hotels/{hotelId}/requests/summary": {
      "get": {
        "operationId": "HotelRequestsController_getRequestsSummary",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guest request summary"
          }
        },
        "tags": [
          "hotel-requests"
        ]
      }
    },
    "/hotels/{hotelId}/requests/{requestId}": {
      "get": {
        "operationId": "HotelRequestsController_getRequestDetail",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đã lấy yêu cầu của khách"
          }
        },
        "tags": [
          "hotel-requests"
        ]
      }
    },
    "/hotels/{hotelId}/requests/{requestId}/status": {
      "patch": {
        "operationId": "HotelRequestsController_updateRequestStatus",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật yêu cầu của khách"
          }
        },
        "tags": [
          "hotel-requests"
        ]
      }
    },
    "/hotels/{hotelId}/requests/{requestId}/assignment": {
      "patch": {
        "operationId": "HotelRequestsController_updateRequestAssignment",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật phân công yêu cầu của khách"
          }
        },
        "tags": [
          "hotel-requests"
        ]
      }
    },
    "/hotels/{hotelId}/requests/{requestId}/events": {
      "post": {
        "operationId": "HotelRequestsController_createRequestEvent",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đã tạo sự kiện yêu cầu của khách"
          }
        },
        "tags": [
          "hotel-requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "VietSage Auth Service API",
    "description": "HTTP API contract for auth, rbac, and hotel-users modules",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}